requirements.txt
Flask>=2.3.0
Flask-CORS>=4.0.0
torch>=2.0.0
torchvision>=0.15.0
onnxruntime>=1.20.0
Pillow>=10.0.0
numpy>=1.24.0
requests>=2.31.0 